chore_: move InitFilters to separate file
